refactor(user): BaseEntity 통합 — common/entity → common/core 단일화
- user/user-role 도메인이 쓰던 구형 BaseEntity(common/entity/base.entity.ts, id 없음, uuid 없음)를 신형 BaseEntity(common/core/base.entity.ts)로 교체
- 중복 선언된 @PrimaryGeneratedColumn() id 제거 (BaseEntity에서 제공)
- users/user_roles 테이블에 uuid, 커서 페이지네이션용 인덱스([createdAt, id])가 자동 추가됨
- common/entity/base.entity.ts 삭제 (다른 사용처 없음)
- synchronize 환경에서 자동 반영됨. 프로덕션은 별도 마이그레이션 필요
